requirements.txt
Django==2.1.5
Pillow==5.4.1
pytz==2018.9
django-logentry-admin==1.0.4
sqlparse==0.2.4
qrcode==6.1
django-sql-explorer==1.1.2
django-iprestrict==1.7.0
django-generic-json-views==0.8
django-admin-rangefilter==0.3.11
pyotp==2.2.7
codicefiscale==0.9
